backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
email-validator
pydantic
pydantic-settings

